Updated: single script generator method

+++ includes/traits/Generator.php
@@ -5,7 +5,8 @@ if (!defined('ABSPATH')) {
     exit;
 } // Exit if accessed directly
 
-use MatthiasMullie\Minify;
+use \MatthiasMullie\Minify;
+use \Elementor\Plugin;
 
 trait Generator
 {
@@ -63,6 +64,19 @@ trait Generator
     ];
 
     /**
+     * Collect elements in a page or post
+     *
+     * @since 3.0.0
+     */
+    public function collect_elements($widget) {
+        $this->transient_elements[] = $widget->get_name();
+    }
+
+    public function set_transient_status($post_id) {
+        update_post_meta($post_id, 'eael_has_transient_elements', true);
+    }
+    
+     /**
      * Collect dependencies for modules
      *
      * @since 3.0.0
@@ -79,86 +93,6 @@ trait Generator
         }
 
         return array_unique($paths);
-    }
-
-    /**
-     * Search elements in a post.
-     *
-     * @since 3.0.0
-     */
-    public function widgets_in_post($post_id)
-    {
-        $elements = array();
-        $sections = json_decode((string) get_post_meta($post_id, '_elementor_data', true));
-
-        if (empty($sections)) {
-            return $elements;
-        }
-
-        foreach ((array) $sections as $section) {
-            if (empty($section->elements)) {
-                continue;
-            }
-            foreach ((array) $section->elements as $element) {
-                if (empty($element->elements)) {
-                    continue;
-                }
-                foreach ((array) $element->elements as $widget) {
-                    if (@$widget->widgetType) {
-                        $elements[] = $widget->widgetType;
-                    } else {
-                        foreach ((array) $widget as $inner_section) {
-                            if (empty($inner_section)) {
-                                continue;
-                            }
-
-                            foreach ((array) $inner_section as $inner_elements) {
-                                if (empty($inner_elements->elements)) {
-                                    continue;
-                                }
-                                foreach ((array) $inner_elements->elements as $inner_widget) {
-                                    if (@$inner_widget->widgetType) {
-                                        $elements[] = $inner_widget->widgetType;
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        }
-
-        $elements = array_map(function ($val) {
-            $val = str_replace(['eael-'], [''], $val);
-
-            return str_replace([
-                'eicon-woocommerce',
-                'countdown',
-                'creative-button',
-                'team-member',
-                'testimonial',
-                'weform',
-                'cta-box',
-                'dual-color-header',
-                'pricing-table',
-                'filterable-gallery',
-            ], [
-                'product-grid',
-                'count-down',
-                'creative-btn',
-                'team-members',
-                'testimonials',
-                'weforms',
-                'call-to-action',
-                'dual-header',
-                'price-table',
-                'filter-gallery',
-            ], $val);
-        }, $elements);
-
-        $elements = array_intersect(array_keys($this->registered_elements), $elements);
-
-        return $elements;
     }
 
     /**
@@ -211,29 +145,78 @@ trait Generator
     }
 
     /**
+     * Check if cache files exists
+     *
+     * @since 3.0.0
+     */
+    public function has_cache_files($post_id = null)
+    {
+        $css_path = EAEL_ASSET_PATH . DIRECTORY_SEPARATOR . ($post_id ? 'eael-' . $post_id : 'eael') . '.min.css';
+        $js_path = EAEL_ASSET_PATH . DIRECTORY_SEPARATOR . ($post_id ? 'eael-' . $post_id : 'eael') . '.min.js';
+
+        if (is_readable($css_path) && is_readable($js_path)) {
+            return true;
+        }
+
+        return false;
+    }
+
+    /**
      * Generate single post scripts
      *
      * @since 3.0.0
      */
-    public function generate_post_scripts($post_id, $elements = null)
+    public function generate_post_scripts($query)
     {
-        if (!is_array($elements)) {
-            $elements = $this->widgets_in_post($post_id);
+        if (Plugin::$instance->preview->is_preview_mode()) {
+            return;
         }
 
-        if (empty($elements)) {
-            $css_path = EAEL_ASSET_PATH . DIRECTORY_SEPARATOR . 'eael-' . $post_id . '.min.css';
-            $js_path = EAEL_ASSET_PATH . DIRECTORY_SEPARATOR . 'eael-' . $post_id . '.min.js';
+        if (get_post_meta($query->queried_object_id, 'eael_has_transient_elements', true) || !$this->has_cache_files($query->queried_object_id)) {
+            $elements = array_map(function ($val) {
+                $val = str_replace(['eael-'], [''], $val);
 
-            if (file_exists($css_path)) {
-                unlink($css_path);
-            }
+                return str_replace([
+                    'eicon-woocommerce',
+                    'countdown',
+                    'creative-button',
+                    'team-member',
+                    'testimonial',
+                    'weform',
+                    'cta-box',
+                    'dual-color-header',
+                    'pricing-table',
+                    'filterable-gallery',
+                ], [
+                    'product-grid',
+                    'count-down',
+                    'creative-btn',
+                    'team-members',
+                    'testimonials',
+                    'weforms',
+                    'call-to-action',
+                    'dual-header',
+                    'price-table',
+                    'filter-gallery',
+                ], $val);
+            }, $this->transient_elements);
 
-            if (file_exists($js_path)) {
-                unlink($js_path);
+            $elements = array_intersect(array_keys($this->registered_elements), $elements);
+
+            if (empty($elements)) {
+                $css_path = EAEL_ASSET_PATH . DIRECTORY_SEPARATOR . 'eael-' . $query->queried_object_id . '.min.css';
+                $js_path = EAEL_ASSET_PATH . DIRECTORY_SEPARATOR . 'eael-' . $query->queried_object_id . '.min.js';
+
+                if (file_exists($css_path)) {
+                    unlink($css_path);
+                }
+
+                if (file_exists($js_path)) {
+                    unlink($js_path);
+                }
+            } else {
+                $this->generate_scripts($elements, 'eael-' . $query->queried_object_id);
             }
-        } else {
-            $this->generate_scripts($elements, 'eael-' . $post_id);
         }
     }
 }
